fixed dialyzer errors in async_iterator_move
Hi I fixed two problems with the specs of async_iterator_move. That said please double check of this new specs are correct, they make sense from the way the functions is called and from looking at the C-Code it also seems to be the proper return but I'm not that deep into the levedb C-code that I feel 100% confident about reading it correctly.