Use system snappy lib
We were shipping an old version of snappy, storing it as a tar file,
extracting, building the lib and then static compiling it into our
eleveldb.so.

Why do this? Snappy is a library (snappy) via brew on OSX, via apt-get/
yum (libsnappy-dev).

Our version is also old, and reduces farm to plate traceability.